@@ -4,14 +4,13 @@ using Verse;
 // ReSharper disable RedundantAssignment
 // ReSharper disable InconsistentNaming
 
-namespace EveryoneCanWork
+namespace EveryoneCanWork;
+
+[HarmonyPatch(typeof(LifeStageWorkSettings), "IsDisabled")]
+public class LifeStageWorkSettings_IsDisabled_Patch
 {
-    [HarmonyPatch(typeof(LifeStageWorkSettings), "IsDisabled")]
-    public class LifeStageWorkSettings_IsDisabled_Patch
+    private static void Postfix(ref bool __result)
     {
-        private static void Postfix(ref bool __result)
-        {
-            __result = false;
-        }
+        __result = false;
     }
 }
